Add "Chimaps in a few lines of code" notebook
This directly copies most of the content from
https://topotoolbox.wordpress.com/2017/08/18/chimaps-in-a-few-lines-of-code-final/
with some modifications for the functionality that currently exists in
pytopotoolbox. This is not expected to work completely (see #73 for
details).

It does depend on the Mendocino data, which should be added to
the DEMs repository if we want to use it.