Remove version specifications from pyproject.toml
I changed my mind about specifying versions in pyproject.toml (#68),
but I wanted to go ahead and merge that one. I think having
requirements.txt for developers and the automated testing
infrastructure to have a consistent environment is all right, though
we might need to think about how we decide when to bump those
versions. We shouldn't constrain users of the library to also have a
particular version of the dependencies unless it is necessary for our
topotoolbox functionality that a certain dependency version be
available.