Add a release workflow
This moves towards a solution for #67.

This workflow uses cibuildwheel to build binary wheels of
pytopotoolbox on various architectures. When a release is made it will
build and test all of these wheels and upload them as assets to the
release.

It also sets the CMAKE_POSITION_INDEPENDENT_CODE flag because it seems
like we need position independent code in our static libtopotoolbox
library to link it successfully on certain platforms.