Restore NaNs in the original DEM after filling sinks (#88)
This ensures that the original DEM is not changed after calling
`fillsinks`. This is not a problem in `GridObject.fillsinks`, because
we make a copy of the DEM with the `astype` call while no copy is made
in `FlowObject`. We probably should not make that copy if we can avoid
it, though, so this change handles things correctly in both cases.