Add reproject method
Resolves #82

`GridObject.reproject` feeds `rasterio.warp.reproject` with the right
source information from the provided `GridObject` and with the desired
CRS and resolution. This requires that a user pass a
`rasterio.crs.CRS` object and possibly a resolution. It does not
automatically find the CRS corresponding to the UTM zone of the
GridObject or anything, which MATLAB's `reproject2utm` does. We might
consider adding a separate method to do so or by dispatching on the
crs argument, i.e. DEM.reproject(crs='utm') would construct the CRS
for the UTM zone of the DEM, but
`DEM.reproject(crs=rasterio.CRS.from_epsg(32613)` would use the
provided CRS object.