Make shape, rows and columns properties of GridObject (#93)
See the discussion in #90.

This ensures that these attributes remain synchronized with the
underlying Numpy array that stores the data, but should not change the
API of the object. It will break code that directly sets these
attributes. I believe I have found and corrected all of those in
the package.