Fix links to CONTRIBUTING.md in docs (#94)
The CONTRIBUTING.md was not being picked up because we didn't have `myst_parser` installed in our docs environment. I added `myst_parser` to `docs/conf.py` and `docs/requirements.txt`, and I moved CONTRIBUTING.md to the docs folder for ease of access. The link on the README has also been updated.