Use `fillsinks_hybrid` (#99)
This adds a `hybrid` option to `GridObject.fillsinks` and `FlowObject` that dispatches to the hybrid morphological reconstruction. My experiments suggest that this is almost always faster than the sequential reconstruction, so `hybrid=True` is the new default, which chooses `fillsinks_hybrid`.

A test is added that ensures that the results from the two algorithms are the same.