Move tutorial notebook to tutorial directory
The tutorial directory now contains a series of notebooks modeled on
MATLAB's "Getting Started with TopoToolbox"
guide. tutorial/tutorial.ipynb serves as the root of this toctree,
with the links to the next chapters in the tutorial at the
bottom. tutorial/grid.ipynb, tutorial/flow.ipynb, and
tutorial/stream.ipynb are currently placeholders into which existing
documentation can be moved.