Add Curvature to GridObject (#100)
Curvature is numerically unstable and very sensitive both to the data and to numerical errors. There are numerical differences between this implementation and the MATLAB one especially when the denominator of the calculation is close to zero (i.e. when the gradient is zero). We should be careful about making comparisons between the two. The MATLAB implementation sets the curvature to zero when the denominator is exactly zero. This implementation instead adds a small offset to the denominator.