Create a Getting Started with TopoToolbox guide (#102)
This set of changes expands the existing Tutorial to include the topics covered in the "Getting Started with TopoToolbox" section of the MATLAB documentation. All of these notebooks are in the docs/tutorial directory, and they form their own subtree of the table of contents. I deleted a few of the existing examples that overlapped with the expanded tutorial.

For the most part I copied the text from the MATLAB documentation and ported the code over.