Implement `StreamObject.chitransform` (#115)
This commit implements the chi transform for StreamObject, resolving #75. It uses streamquad_trapz from libtopotoolbox to perform the numerical integration. Everything else sets up the data to be integrated.

There is a new StreamObject.distance method that returns the interpixel distance for each edge in the stream network. This is not identical to the STREAMobj.distance method that computes upstream distance in MATLAB. It is needed to compute the integration weights.

StreamObject.ezgetnal extracts node attributes from a GridObject or an ndarray. It does not do all of the shape and projection validation that STREAMobj.ezgetnal does, so be careful not to pass unaligned grids.

This has been tested manually on the perfectworld DEM, where it agrees with MATLAB's STREAMobj.chitransform with some floating-point error. Fully automated testing on real DEMs is slightly complicated because libtopotoolbox-based flow routing does not currently give identical answers when there are flat areas. However, it should be possible to test the chitransform itself by constructing a StreamObject directly and verifying that it gives the expected result.

Stride information from the GridObject's underlying array is now copied to the FlowObject and StreamObject. This is to simplify writing
memory-layout-agnostic indexing functions for the FlowObject and
StreamObject. In particular, StreamObject.distance needs the strides to determine which pixels are 4-adjacent to a central pixel.