Add Drainage basin delineation section to the Getting Started guide
This copies the text from the MATLAB Getting Started guide, but adds an
explanation for `shufflelabel` and uses a different color scheme to
plot the drainage basins.

The resulting drainage basins look a little weird, but that is because
of our known problems routing flow over flat
regions (TopoToolbox/libtopotoolbox#122).