Make fillsinks memory-order agnostic (#125)
This change represents one way that we might make our interface with libtopotoolbox agnostic to the memory order of our underlying numpy arrays. I have added a property dims to GridObject, which reverses the order of the dimensions if the array is in row-major layout. GridObject.shape still returns (rows, columns). If we pass self.dims instead of self.shape to the fillsinks wrapper, we will now get the same, correct, answer if the GridObject.z array is column- or row-major. There is also a test that confirms this behavior.