Add test dependencies as an optional dependency (#126)
The changes in #125 use opensimplex directly in the tests, and I
wanted to be clear that you need opensimplex to run the tests. I left
the "opensimplex" optional dependency in there to avoid breaking
workflows and because it does provide functionality that we use
outside of the tests.

I also added pytest to the "test" dependency. I believe this is the only
other package we use directly. All of the other test dependencies are
transitive dependencies. Cleanly installing everything using `pip
install .[test]` in a new virtual environment works, and the tests are
able to run successfully.