Add xy method to StreamObject
`StreamObject.xy` return a list of lists of tuples giving the
coordinates of nodes in the stream network that are connected into
continuous stream segments. It constructs this list by conducting a
depth-first search on the graph, appending nodes to a segment until
the stream reaches an outlet or a sink or a previously visited
node. The edge list representation does not allow a depth-first search
on its own, so an adjacency list is constructed first.

The main difference between `StreamObject.xy` and MATLAB's
STREAMobj2XY is that `StreamObject.xy` returns a list of lists of
coordinates, while MATLAB returns a pair of NaN-punctuated
vectors. The preferred way to plot many lines in matplotlib is with a
`LineCollection`, which takes a list of lists of coordinates, which is
why this representation was chosen. This should make it possible to do
something like

```python
segments = LineCollection(S.xy(), colors='r', linewidths=0.5)
ax.add_collection(segments)
```

to plot the `StreamObject` in the provided axis. I will add this
plotting method in a separate PR.