Replace GridObject.show with GridObject.plot (#142)
* Replace show with plot

This commit follows the recent change to the StreamObject plotting
interface (#130) by adding a `plot` method to GridObject that plots
the object with imshow within the provided axes.

This enables more composable plotting systems and plays nicely with
the rest of the matplotlib interface. It can, however, make plotting
quite verbose, especially when you need to add a colorbar (see
theexcess topography and magicfunctions notebooks for some
examples). I considered adding an option to add a colorbar to the plot
function, wrapping up all of this axis divider stuff, but there are
too many possible options that we would need to expose to make a nice
interface.

This behavior was technically already available because GridObject is
an array-like object, so it could be passed to imshow directly. This
simply makes plotting more obvious. We should probably provide some
examples of the array-like interface.