Use flow_accumulation_edgelist to compute flow accumulation
We previously used the node/direction version of flow accumulation
from libtopotoolbox, but since we now store the edge list rather than
the full topologically sorted list of nodes, we have to use
`flow_accumulation_edgelist`. This makes the change in both FlowObject
and StreamObject, which uses the raw `_flow.flow_accumulation` function.

We also have to pass a `fraction` argument. This is a pointless array
allocation, but it is needed by the interface for
`flow_accumulation_edgelist`.