Use the new libtopotoolbox flow routing API (#147)
This should resolve #146. It required changing flow_routing_targets to flow_routing_d8_edgelist as expected, but it also required us to switch from using libtopotoolbox's flow_accumulation to flow_accumulation_edgelist. I had hoped to make this switch otherwise, so this was a good opportunity to do it. There was one other minor libtopotoolbox change to drainagebasins that this effort revealed (TopoToolbox/libtopotoolbox#169).

The changes to StreamObject were minimal, updating it to use the new flow accumulation interface and cleaning up some indices in the constructor function. The normal StreamObject functions already use proper edge lists, so once the constructor was fixed, everything worked properly.

The tests did catch a lot of the necessary changes. I added an index check to test_flow_object.py make sure that we don't include the -1 edges in the FlowObject edge list.