Add plotdz to StreamObject
Resolves #114

This was fairly straightforward to implement using the same pattern as
normally plotting a StreamObject with a LineCollection. A small
modification to `StreamObject.xy` allows a user to pass a tuple of x,
and y node attribute lists that are segmented as appropriate. The
segmentation is separate from the values of the generated points. If
no data are passed it, the previous behavior, which returns the x and
y indices of the pixels in the GridObject is preserved.

We use libtopotoolbox's `streamquad_trapz_f32` to compute upstream
distance though there are simpler ways to do so with other traversal
methods, and we should change that once such traversals are available
in libtopotoolbox.

The optional arguments supported by MATLAB's plotdz are not all
supported yet. Passing `linewidths` as a keyword argument will work,
because this is forwarded to the underlying LineCollection. `dunit`
and `doffset` should be relatively straightforward to add. Plotting
multicolored line segments with LineCollection is possible, but rather
complicated, and I have not yet figured out a good way to handle this.