Implement klargestconncomps
This version of `klargestconncomps` uses traversals implemented in
libtopotoolbox to compute the size of each connected component of the
stream network and to propagate the identities of the largest
connected components upstream.

test/stream_object.py adds some tests that ensure that `trunk` and
`klargestconncomps` create subgraphs of the original stream network. A
helper function `issubgraph` could also be included in the
`StreamObject` interface, but it is not particularly efficient as it
relies on set operations.