Add hillshade to GridObject
This adds a wrapper for the libtopotoolbox hillshade function and a
method to GridObject that calls that wrapper. The procedure for
converting the azimuth angle from a geographic bearing into the grid's
coordinate system is rather involved, but it seems to work for our
example DEMs. A test is added that compares the hillshades of random
DEMs in row- and column-major order to test that this azimuth angle
computation is correct. Some comments are added to help explain the
process of rotating the azimuth angle.

Note that `GridObject.hillshade` only computes the hillshade, it does
not display it. This behaves differently from TopoToolbox v2, but a
separate plotting function or a `hillshade` argument to
`GridObject.plot` fits more naturally with our existing visualization
API.