Only run the release workflow on the main TopoToolbox repository
All of the jobs in the release workflow now check that they are
running from within the TopoToolbox/pytopotoolbox repository and will
not run on forks. Because the job had to be triggered by a release,
and because our PyPI trusted publishing setup should only work from
the TopoToolbox repository, I think this is not entirely necessary,
but I have made the same change to the release workflows in
topotoolbox3 and libtopotoolbox.