Implement subgraph (#177)
This adds a method `subgraph` to `StreamObject` that extracts subsets
of the stream network by reindexing the nodes of the StreamObject.

This code duplicates that in `trunk` and in `klargestconncomps` and
those sites should be updated to call `subgraph`.

MATLAB TopoToolbox uses `clean` to remove singleton nodes of the
stream network and provides the option to return indices into the
original stream network, both of which might be useful to add in the
future. TODOs have been added there.

A test is added to tests/test_stream_object.py. This test constructs
subgraphs of the Wide DEM and Tall DEM stream networks using
downstream distance thresholds and tests properties of the resulting
subgraphs.