Implement `FlowObject.flowpathextract` in the downstream direction
See #189

A new method is added to FlowObject in src/topotoolbox/flow_object.py
that computes the indices of a flow path in the downstream direction
using the downstream Boolean semiring traversal followed by some index
extraction.

A test is added to tests/test_flow_object.py that ensures that the
flow path extracted here is identical to the StreamObject constructed
using the `channelheads` argument.