Plot GridObjects and StreamObjects with coordinates
Resolves #187

This commit adds an `extent` property to `GridObject`, which
constructs a bounding box in the order that matplotlib expects for its
`extent` argument to `imshow`. This bounding box is then supplied to
`imshow` to plot the `GridObject` with the proper coordinates.

`StreamObject.xy` now defaults to using the geotransform to convert
the pixel indices into coordinates. `StreamObject.plot` therefore now
plots into the coordinate space. The other use of `StreamObject.xy` in
`StreamObject.plotdz` uses the `data` argument to `plotdz`, so it is
not affected.