Implement imposemin
imposemin can take either a FlowObject or a StreamObject, so it is
implemented as a standalone function in the new file
src/topotoolbox/stream_functions.py. It uses the (min,+) traversal
implemented in libtopotoolbox.

imposemin for a FlowObject requires the interpixel distances. I have
added a method very similar to StreamObject.distance that computes
these.

Tests are added for both the StreamObject and FlowObject versions. The
tests ensure

1. imposemin does not modify the original DEM. Certain copies are made
explicitly in the implementation of imposemin to make sure that this
is true.
2. The carved DEM is not above the original
3. The gradient is constrained within numerical error by the
minimum_slope argument.