Correct the hillshade order test
This test randomly generated double-precision DEMs, which were then
passed to libtopotoolbox's hillshade as single-precision DEMs. The
results were then incorrect, but compared equal to one another. This
change forces the DEMs to be single-precision.

However, this change causes the test to fail because the geotransform
created for the column-major array is identical to that of the
row-major array. A rotation is applied to the column-major
geotransform to correct this. `np.allclose` is now used for the
comparison because the separate geotransforms lead to small numerical
differences between the two hillshades. The length scale of the random
DEMs is also increased because unrealistically steep gradients lead to
increased numerical differences in the hillshades.