Fix the row/column major test for excesstopography (#207)
First we apply the datatype fix previously needed for fillsinks (#206)
and hillshade (#205). Then we set the cellsize to make sure that
gradients are computed correctly in excesstopography. This corrects
the test, but the test now fails.

The following changes make the tests pass

1. Use the newly available (#196) np.asarray functionality for
GridObjects to convert the threshold slope and DEM GridObjects.
2. Use np.full_like to construct `threshold_slopes` so that it matches
the datatype, shape and order of the DEM. (#201)
3. Remove the datatype and memory layout checks.
4. Pass dims instead of shape (#200)