Implement `validate_alignment` (#211)
Resolves #209

This function applies to all three Topotoolbox objects, so it is put
into the `topotoolbox.utils` module. This may not be the best place
for it in the long run.

`validate_alignment` checks whether two objects have the same shape
and coordinate system. It does not check the strides of the object,
because we do not currently expose the strides of `GridObject`, but
that can be added in the future.

These tests make sure that topotoolbox functions copy the appropriate
metadata into their outputs.