Use ezgetnal and the __array__ interface in imposemin
`ezgetnal` extracts the data relevant to the Flow-/StreamObject, and
`np.asarray` converts a GridObject to an array, and is a no-op
otherwise. This will throw an error on Numpy >= 2.0 if it is unable to
create the copy because the data type is incorrect. This should not
work, however, on Numpy < 2.0.

A test is added to make sure that the error gets thrown if the data
type is incorrect.