Add dtype argument to ezgetnal
Since we are copying the array anyway, it is not so hard to make sure
that the returned arrays have a given element type. This will also be
useful in `imposemin` to make sure that extracted node attribute lists
are safe to pass to libtopotoolbox

Tests are added to ensure that the dtype is changed as expected.