Return copies from ezgetnal (#220)
This PR changes ezgetnal so that it always returns a copy of the input array. This problem emerged when working on imposemin (#194) in trying to ensure that we pass views of the appropriate arrays to libtopotoolbox. This does duplicate arrays in some places where it isn't strictly necessary, and we might consider adding a copy argument akin to that on numpy.array that can suppress these copies.

Since we are copying the array anyway, it is not so hard to make sure that the returned arrays have a given element type especially with #219.

Tests are added to make sure that the copying works as anticipated and that the dtype is changed when requsted.

This also changes the implementation of the __array__ method on GridObject so that it only copies when copy=True. Otherwise it trusts that Numpy will figure things out when you call numpy.array(dem, copy=False). Making this compatible with the numpy pre-2.0 behavior while also relying on it to throw errors when copies were not possible was finicky.