Use the copying ezgetnal and dtype argument in imposemin
s.ezgetnal(dem, dtype=np.float32) should create the right kind of
object (ndarray or GridObject) with the right data
type (np.float32). This will now work with a float64 input array, so
the test has been changed to make this work. However, it will convert
the input to single precision before running, and the returned object
will be a float32. This may be slightly unexpected, so it is
documented.