Call np.array on the z member rather than the GridObject
This is a subtle interaction between the array interface and the
memory layout of arrays. The wide_dem is a column-major array, but
calling `z = np.array(wide_dem, dtype=np.float64)` creates a row-major
array by default because the memory layout of the `GridObject` does
not get passed along. This is fixed in this test by calling
`np.array(wide_dem.z, dtype=np.float64)`.

This needs further investigation (#198), but it is not specific to
`imposemin`.