Implement imposemin (#194)
imposemin can take either a FlowObject or a StreamObject, so it is
implemented as a standalone function in the new file
src/topotoolbox/stream_functions.py. It uses the (min,+) traversal
implemented in libtopotoolbox. The version of libtopotoolbox has
been incremented to include traverse_down_f32_min_add.

imposemin for a FlowObject requires the interpixel distances. I have
added a method very similar to StreamObject.distance that computes
these.

Tests are added for both the StreamObject and FlowObject versions. The
tests ensure

1. imposemin does not modify the original DEM. Certain copies are made
explicitly in the implementation of imposemin to make sure that this
is true.
2. The carved DEM is not above the original
3. The gradient is constrained within numerical error by the
minimum_slope argument.

Furthermore, a test is added to ensure that the function still works
if the supplied DEM has an element type other than single precision
float. The input will be converted to float32 before being passed to
libtopotoolbox, and the output will be float32, so some loss of precision
is possible.

The stream_functions module is added to the documentation.

There is a subtle interaction between the recently implemented array interface and the
memory layout of arrays that is revealed in the FlowObject.imposemin test. The wide_dem is a column-major array, but
calling `z = np.array(wide_dem, dtype=np.float64)` creates a row-major
array by default because the memory layout of the `GridObject` does
not get passed along. This is fixed in this test by calling
`np.array(wide_dem.z, dtype=np.float64)`. This needs further investigation (#198), but it is not specific to `imposemin`.