Add memory ordering test for identifyflats
As with the other tests, we create a row-major and an identical
column-major array, pass both to identifyflats and compare the
outputs. We also check that the memory layout is preserved by
identifyflats.

This test will fail because the memory layout is not preserved by
identifyflats.

Test the memory layout for flats/sills output as well as raw output