Add memory ordering test for identifyflats (#221)
See #199

As with the other tests, we create a row-major and an identical column-major array, pass both to identifyflats and compare the outputs. We also check that the memory layout is preserved by identifyflats.

The row-major and column-major arrays already produced identical results in identifyflats because the input was converted to column-major before passing it through. This was caught by the second set of tests. By using self.dims instead of self.shape (#200) and removing the explicit conversion (#201), this test passes.

This also changes the return type of the raw=True option. Instead of returning tuple(output_grid), it returns (output_grid,). The former is a tuple of the rows of the output grid, and the latter is a one-element tuple of the 2D output grid, which is what we should use instead. Uses of identifyflats are updated for this new return type.