Type checking changes for Numpy v2.2 (#222)
Numpy v2.2 has introduced some new type checking annotations, and
while we don't rely on any of these annotations, mypy unfortunately
labels some errors.

Most importantly, there are new shape types, but not all of the numpy
operations respect shape types, and these return a default shape type
that is incompatible with a more specific type. If you then try to
assign this to a variable that has already been assigned, you'll get a
type error in mypy.

This affects us because we assign an empty ndarray to GridObject.z
when we create a new GridObject from scratch. So calling

output = GridObject()
output.z = np.zeros_like(dem.z)

will result in a type error, because the empty array created for
output.z has a shape type that is incompatible with the type of
np.zeros_like(dem.z).

This is fixed for now by annotating self.z with the most general
NDArray type when we create an empty GridObject.

There is also a type problem when multiplying a float32 array by
180.0/np.pi in gradient8, which is fixed using np.rad2deg instead.