Fix memory ordering problems with FlowObject (#230)
Due to limitations in libtopotoolbox, flow routing is sensitive to the
memory order of the input arrays. This has not been fixed, but
TopoToolbox/libtopotoolbox#186 modifies the API for
flow_routing_d8_carve and flow_routing_d8_edgelist to take an order
parameter (0 => column-major, 1 => row-major). This parameter corrects
the search for flow directions so that the flow routing should be
identical for identical arrays in either memory order. Our
CMakeLists.txt is updated to include this change.

We do not pass the order flag through the Python API, but instead
compute it by examining the flags() method on the direction
pybind11::array. Note that all of the arrays should have the same
memory ordering, but we do not currently enforce that in any way.

A test is added that compares the FlowObject for identical row-major
and column-major arrays, following #199. The changes required for #200
and #201 are implemented in the FlowObject constructor.