Add automatic downloads from OpenTopography  (#216)
This adds two new functions to utils.py:

    load_open_topography()
    read_from_cache()
Closes #215

This adds the ability to Load DEM's from OpenTopography by specifying the bounds, global raster dataset and an API-Key. These will be cached to avoid having to download them again if called with the same arguments. There is also the option to save them at a desired path. To make it easier to load already cached datasets, the function read_from_cache() can get them from the cache. This way the user can avoid having to input the whole path, if they haven't saved the TIF at a different place using the save_path argument.