Fix type stability bug in ezgetnal and plotdz (#231)
* Ensure that ezgetnal does not change type of array by default

nal.astype(None, copy=False) does not simply return nal, but converts
it to float64. If dtype is None, we pass nal.dtype and the astype
statement does nothing. Otherwise, we pass dtype, and it converts the
array as desired.

* plotdz: pass single arrays to streamquad_trapz_f32

The z array doesn't have to be float32, because it is just used in
plotting, but the dist and a arrays are passed to libtopotoolbox and
must be single precision floats.

* Test that ezgetnal does not change data type