Add upstreamto and downstreamto methods on StreamObject
This is part of an effort to implement
`StreamObject.modify` (#140). The basic idea is to implement the
different options of `modify` as small methods on StreamObject and
then either to have `modify` simply dispatch to these methods or to
forgo a separate `modify` function entirely and expect users to use
the smaller modification methods.