Implement `StreamObject.upstreamto`, `StreamObject.downstreamto` and `StreamObject.clean` (#242)
* Wrap `traverse_up_u32_or_and`

This replaces `traverse_up_u32_and`, which is not currently used
anywhere (it was planned for the implementation of trunk, which is
done in python). `traverse_up_u32_and` has been removed from
libtopotoolbox and replaced with `traverse_up_u32_or_and`.

* Add upstreamto and downstreamto methods on StreamObject

This is part of an effort to implement
`StreamObject.modify` (#140). The basic idea is to implement the
different options of `modify` as small methods on StreamObject and
then either to have `modify` simply dispatch to these methods or to
forgo a separate `modify` function entirely and expect users to use
the smaller modification methods.

* Implement StreamObject.clean

This is a subgraph based on the degree of each node. This is needed to
properly test that upstreamto from the outlets is identical to the
original stream network.