Add memory order test for flow_accumulation
Create DEMs in row- and column-major order, route flow and compute
flow accumulation. These two outputs should be identical.

This test fails. The flow accumulation rasters are generated in
column-major order, but the source and target indices are either row-
or column-major depending on the source DEM, so the row-major accesses
the incorrect values. This is incorrect, but does not access memory
out of bounds.