Add memory order test for flow_accumulation (#245)
* Add memory order test for flow_accumulation

Create DEMs in row- and column-major order, route flow and compute
flow accumulation. These two outputs should be identical.

This test fails. The flow accumulation rasters are generated in
column-major order, but the source and target indices are either row-
or column-major depending on the source DEM, so the row-major accesses
the incorrect values. This is incorrect, but does not access memory
out of bounds.

* Fix memory order test for flow_accumulation

This is a minimal solution to the failed memory order test. It adds an
`order` attribute to the `FlowObject` and creates the accumulation and
weights rasters with the correct order. The test now passes. I am
unsure if this is a good long-term solution.

Both `FlowObject` and `StreamObject` need to know something about the
memory order of their underlying grids because they need to be able to
unravel their source and target to index into the grid. For
`FlowObject`

  grid.z[np.unravel_index(fd.source, fd.shape, order=fd.order)]

while for `StreamObject`, you need the extra indirection of the
`stream` array

  grid.z[np.unravel_index(s.stream[s.source], s.shape, order=s.order)]

Passing around the order as a string flag feels unsafe, but it does
solve this problem.

* Set the type of self.order

* Fix error in condition for FlowObject.order