Test the memory order independence of drainagebasins  (#246)
See #199

This test requires some fiddly indexing logic to deal with
permutations of the drainage basin labels, which I hope is
well-explained by the comments.

This test fails at first, because the basins array is created with a column-major order.
Fixing it is mostly a matter of using FlowObject.order to construct the
right kind of output arrays. It also requires a new FlowObject.dims
attribute to pass the dimensions in the correct order.