drainagebasins: Use the u32 or-and upstream traversal with outlets (#250)
If the specified outlets were not outlets of the FlowObject, zeros
would be propagated upstream and overwrite the basins. The MATLAB code
checks whether the target is zero before overwriting, but we can also
use the or-and upstream traversal as shown here. The edge weights are
masks with every bit set, and the node attribute list is the drainage
basin labels, so this assigned drainage basin labels upstream, but
won't overwrite with zeros. Weirder things will happen if you have
multiple flow directions: you'll end up with basin labels that are
bitwise mixtures of the two downstream labels and thus that might not
be a valid basin label. But drainage basins are not defined for
divergent flows anyway, so this shouldn't be a problem.