Implement `GridObject.plot_surface` (#253)
Resolves #135 and #133

This PR adds a plot_surface method to GridObject. This is a very basic interface to mplot3d's plot_surface. As with many of our matplotlib functions, there is a lot of complexity buried in there, so creating a 3D surface plot that looks nice is nontrivial. I added a brief example to both the function documentation and the Getting Started Guide that removes some of the more egregious default layout options, namely the aspect ratio and the spacing of the ticks in the z dimension.